feat: add shared model types and refine prisma typing
