+++ packages/platform-core/src/db.d.ts
@@ -1,8 +1,2 @@
 import { PrismaClient } from "@prisma/client";
-export declare const prisma: PrismaClient<{
-    datasources: {
-        db: {
-            url: string;
-        };
-    };
-}, never, import("@prisma/client/runtime/library").DefaultArgs>;
+export declare const prisma: PrismaClient;
