+++ packages/platform-core/src/subscriptionUsage.d.ts
@@ -1,10 +1,4 @@
 import "server-only";
-export interface SubscriptionUsage {
-    id: string;
-    shop: string;
-    customerId: string;
-    month: string;
-    shipments: number;
-}
+import type { SubscriptionUsage } from "@acme/types";
 export declare function getSubscriptionUsage(shop: string, customerId: string, month: string): Promise<SubscriptionUsage | null>;
 export declare function incrementSubscriptionUsage(shop: string, customerId: string, month: string, count?: number): Promise<void>;
