+++ packages/types/src/CustomerProfile.ts
@@ -1,0 +1,5 @@
+export interface CustomerProfile {
+  customerId: string;
+  name: string;
+  email: string;
+}
