+++ packages/types/src/ReverseLogisticsEvent.d.ts
@@ -1,0 +1,14 @@
+export type ReverseLogisticsEventName =
+  | "received"
+  | "cleaning"
+  | "repair"
+  | "qa"
+  | "available";
+
+export interface ReverseLogisticsEvent {
+  id: string;
+  shop: string;
+  sessionId: string;
+  event: ReverseLogisticsEventName;
+  createdAt: string;
+}
