+++ packages/types/src/SubscriptionUsage.ts
@@ -1,0 +1,7 @@
+export interface SubscriptionUsage {
+  id: string;
+  shop: string;
+  customerId: string;
+  month: string;
+  shipments: number;
+}
