+++ packages/types/src/User.ts
@@ -1,0 +1,10 @@
+export interface User {
+  id: string;
+  email: string;
+  passwordHash: string;
+  role: string;
+  resetToken: string | null;
+  resetTokenExpiresAt: Date | null;
+  emailVerified: boolean;
+  stripeSubscriptionId?: string | null;
+}
