+++ packages/types/src/index.d.ts
@@ -19,6 +19,10 @@ export * from "./Shop";
 export * from "./ShopSettings";
 export * from "./Coupon";
 export * from "./SubscriptionPlan";
+export * from "./SubscriptionUsage";
+export * from "./CustomerProfile";
+export * from "./User";
+export * from "./ReverseLogisticsEvent";
 export * from "./Coverage";
 export * from "./upgrade";
 export * from "./ExampleProps";
